--- src/components/control-select/control-select.test.js
@@ -1,91 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import { testCases } from './control-select-test-cases';
-
-describe('ControlSelect', () => {
-  let testCase;
-  let wrapper;
-
-  describe(testCases.basic.description, () => {
-    beforeEach(() => {
-      testCase = testCases.basic;
-      wrapper = shallow(
-        React.createElement(testCase.component, testCase.props)
-      );
-    });
-
-    test('renders', () => {
-      expect(wrapper).toMatchSnapshot();
-    });
-
-    test('onChange gets called with input id and value', () => {
-      const mockEvent = {
-        target: {
-          value: 'foo'
-        }
-      };
-      wrapper.find('select').props().onChange(mockEvent);
-      wrapper.update();
-      expect(wrapper).toMatchSnapshot();
-      expect(testCase.props.onChange).toHaveBeenCalledTimes(1);
-      expect(testCase.props.onChange).toHaveBeenCalledWith('foo', 'testinput');
-    });
-  });
-
-  describe(testCases.disabled.description, () => {
-    beforeEach(() => {
-      testCase = testCases.disabled;
-      wrapper = shallow(
-        React.createElement(testCase.component, testCase.props)
-      );
-    });
-
-    test('renders', () => {
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-
-  describe(testCases.grouped.description, () => {
-    beforeEach(() => {
-      testCase = testCases.grouped;
-      wrapper = shallow(
-        React.createElement(testCase.component, testCase.props)
-      );
-    });
-
-    test('renders', () => {
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-
-  describe(testCases.values.description, () => {
-    beforeEach(() => {
-      testCase = testCases.values;
-      wrapper = shallow(
-        React.createElement(testCase.component, testCase.props)
-      );
-    });
-
-    test('renders', () => {
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-
-  describe(testCases.allOptions.description, () => {
-    beforeEach(() => {
-      testCase = testCases.allOptions;
-      wrapper = shallow(
-        React.createElement(testCase.component, testCase.props)
-      );
-    });
-
-    test('extra props works', () => {
-      const props = wrapper.find('option').last().props();
-      expect(props).toHaveProperty('disabled', true);
-    });
-
-    test('renders', () => {
-      expect(wrapper).toMatchSnapshot();
-    });
-  });
-});
